add 21, 25

+++ Sword2Offer/9_frog_jump_1.cpp
@@ -3,5 +3,5 @@
 int jumpFloorII(int number) {
     if (number <= 0)
         return 0;
-    return pow(2, number - 1);
+    return static_cast<int>(pow(2, number - 1));
 }
